+++ optree/__init__.py
@@ -14,7 +14,7 @@
 # ==============================================================================
 """OpTree: Optimized PyTree Utilities."""
 
-from optree import accessor, integration, typing
+from optree import accessor, functools, integration, typing
 from optree.accessor import (
     DataclassEntry,
     FlattenedEntry,
@@ -27,6 +27,7 @@ from optree.accessor import (
     SequenceEntry,
     StructSequenceEntry,
 )
+from optree.functools import Partial
 from optree.ops import (
     MAX_RECURSION_DEPTH,
     NONE_IS_LEAF,
@@ -93,7 +94,6 @@ from optree.ops import (
 from optree.registry import (
     AttributeKeyPathEntry,
     GetitemKeyPathEntry,
-    Partial,
     register_keypaths,
     register_pytree_node,
     register_pytree_node_class,
@@ -198,10 +198,6 @@ __all__ = [
     'register_pytree_node',
     'register_pytree_node_class',
     'unregister_pytree_node',
-    'Partial',
-    'register_keypaths',
-    'AttributeKeyPathEntry',
-    'GetitemKeyPathEntry',
     # Typing
     'PyTreeSpec',
     'PyTreeDef',