+++ optree/ops.py
@@ -3179,7 +3179,7 @@ def prefix_errors(
     )
 
 
-STANDARD_DICT_TYPES = frozenset({dict, OrderedDict, defaultdict})
+STANDARD_DICT_TYPES: frozenset[type] = frozenset({dict, OrderedDict, defaultdict})
 
 
 # pylint: disable-next=too-many-locals
@@ -3200,8 +3200,7 @@ def _prefix_error(
     prefix_tree_type = type(prefix_tree)
     full_tree_type = type(full_tree)
     both_standard_dict = (
-        prefix_tree_type in STANDARD_DICT_TYPES  # type: ignore[comparison-overlap]
-        and full_tree_type in STANDARD_DICT_TYPES  # type: ignore[comparison-overlap]
+        prefix_tree_type in STANDARD_DICT_TYPES and full_tree_type in STANDARD_DICT_TYPES
     )
     both_deque = prefix_tree_type is deque and full_tree_type is deque  # type: ignore[comparison-overlap]
     if prefix_tree_type is not full_tree_type and (