+++ tests/test_ops.py
@@ -27,6 +27,7 @@ import pytest
 
 import optree
 from helpers import (
+    IS_LEAF_FUNCTIONS,
     LEAVES,
     TREE_ACCESSORS,
     TREE_PATHS,
@@ -37,33 +38,6 @@ from helpers import (
     MyAnotherDict,
     parametrize,
 )
-
-
-def dummy_func(*args, **kwargs):  # pylint: disable=unused-argument
-    return
-
-
-dummy_partial_func = functools.partial(dummy_func, a=1)
-
-
-def is_tuple(tup):
-    return isinstance(tup, tuple)
-
-
-def is_list(lst):
-    return isinstance(lst, list)
-
-
-def is_none(none):
-    return none is None
-
-
-def always(obj):  # pylint: disable=unused-argument
-    return True
-
-
-def never(obj):  # pylint: disable=unused-argument
-    return False
 
 
 def test_max_depth():
@@ -389,13 +363,7 @@ def test_paths_and_accessors(data):
 
 @parametrize(
     tree=TREES,
-    is_leaf=[
-        is_tuple,
-        is_list,
-        is_none,
-        always,
-        never,
-    ],
+    is_leaf=IS_LEAF_FUNCTIONS,
     none_is_leaf=[False, True],
     namespace=['', 'undefined', 'namespace'],
 )
@@ -469,13 +437,7 @@ def test_paths_and_accessors_with_is_leaf(tree, is_leaf, none_is_leaf, namespace
 
 @parametrize(
     tree=TREES,
-    is_leaf=[
-        is_tuple,
-        is_list,
-        is_none,
-        always,
-        never,
-    ],
+    is_leaf=IS_LEAF_FUNCTIONS,
     none_is_leaf=[False, True],
     namespace=['', 'undefined', 'namespace'],
 )
@@ -516,12 +478,7 @@ def test_tree_is_leaf_with_leaves(leaf, none_is_leaf, namespace):
 
 @parametrize(
     tree=TREES,
-    is_leaf=[
-        is_tuple,
-        is_none,
-        always,
-        never,
-    ],
+    is_leaf=IS_LEAF_FUNCTIONS,
     none_is_leaf=[False, True],
     namespace=['', 'undefined', 'namespace'],
 )
@@ -578,12 +535,7 @@ def test_all_leaves_with_leaves(leaf, none_is_leaf, namespace):
 
 @parametrize(
     tree=TREES,
-    is_leaf=[
-        is_tuple,
-        is_none,
-        always,
-        never,
-    ],
+    is_leaf=IS_LEAF_FUNCTIONS,
     none_is_leaf=[False, True],
     namespace=['', 'undefined', 'namespace'],
 )
@@ -1984,43 +1936,3 @@ def test_tree_flatten_one_level(tree, none_is_leaf, namespace):  # noqa: C901
             namespace=namespace,
         )
     ]
-
-
-@parametrize(
-    tree=[
-        optree.Partial(dummy_func),
-        optree.Partial(dummy_func, 1, 2),
-        optree.Partial(dummy_func, x='a'),
-        optree.Partial(dummy_func, 1, 2, 3, x=4, y=5),
-        optree.Partial(dummy_func, 1, None, x=4, y=5, z=None),
-        optree.Partial(dummy_partial_func, 1, 2, 3, x=4, y=5),
-    ],
-    none_is_leaf=[False, True],
-)
-def test_partial_round_trip(tree, none_is_leaf):
-    leaves, treespec = optree.tree_flatten(tree, none_is_leaf=none_is_leaf)
-    actual = optree.tree_unflatten(treespec, leaves)
-    assert actual.func == tree.func
-    assert actual.args == tree.args
-    assert actual.keywords == tree.keywords
-
-
-def test_partial_does_not_merge_with_other_partials():
-    def f(a=None, b=None, c=None):
-        return a, b, c
-
-    g = functools.partial(f, 2)
-    h = optree.Partial(g, 3)
-    assert h.args == (3,)
-    assert g() == (2, None, None)
-    assert h() == (2, 3, None)
-
-
-def test_partial_func_attribute_has_stable_hash():
-    fun = functools.partial(print, 1)
-    p1 = optree.Partial(fun, 2)
-    p2 = optree.Partial(fun, 2)
-    assert p1.func == fun  # pylint: disable=comparison-with-callable
-    assert fun == p1.func  # pylint: disable=comparison-with-callable
-    assert p1.func == p2.func
-    assert hash(p1.func) == hash(p2.func)